pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kinopoisk_django"
version = "0.1.0"
description = "Django REST API project"
authors = [{ name = "Turarbek Sataldiyev", email = "turarbeks@bk.ru" }]
requires-python = ">=3.10"
dependencies = [
    "asgiref==3.10.0",
    "Django==5.2",
    "django-cors-headers==4.3.1",
    "django-extensions==4.1",
    "djangorestframework==3.16.1",
    "djangorestframework_simplejwt==5.5.1",
    "python-decouple",
    "gunicorn==23.0.0",
    "packaging==25.0",
    "pydotplus==2.0.2",
    "PyJWT==2.10.1",
    "pyparsing==3.2.5",
    "python-decouple==3.8",
    "sqlparse==0.5.3",
    "faker"
]
